+++ metadata.rb
@@ -6,7 +6,15 @@ description      'Builds a ruby package using ruby_build and fpm'
 version          '0.3.1'
 
 supports         'debian'
+supports         'linuxmint'
 supports         'ubuntu'
+
+supports         'amazon'
+supports         'centos'
+supports         'fedora'
+supports         'oracle'
+supports         'redhat'
+supports         'scientific'
 
 depends          'fpm-tng', '~> 0.1.6'
 depends          'ruby_build', '~> 0.6'